--- src/main/java/iti/DynamoDBConfig.java
@@ -1,40 +1,0 @@
-package iti;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
-import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRepositories;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.util.StringUtils;
-
-@Configuration
-@EnableDynamoDBRepositories(basePackages = "iti")
-public class DynamoDBConfig {
-
-    @Value("${amazon.dynamodb.endpoint}")
-    private String amazonDynamoDBEndpoint;
-
-    @Value("${amazon.aws.accesskey}")
-    private String awsAccessKey;
-
-    @Value("${amazon.aws.secretkey}")
-    private String awsSecretKey;
-
-    @Bean
-    public AmazonDynamoDB amazonDynamoDB(AWSCredentials awsCredentials) {
-        @SuppressWarnings("deprecation")
-        AmazonDynamoDB amazonDynamoDB = new AmazonDynamoDBClient(awsCredentials);
-        if (!StringUtils.isEmpty(amazonDynamoDBEndpoint)) {
-            amazonDynamoDB.setEndpoint(amazonDynamoDBEndpoint);
-        }
-        return amazonDynamoDB;
-    }
-
-    @Bean
-    public AWSCredentials awsCredentials() {
-        return new BasicAWSCredentials(awsAccessKey, awsSecretKey);
-    }
-}
