+++ src/main/java/iti/db/config/DynamoDBConfig.java
@@ -1,0 +1,47 @@
+package iti.db.config;
+
+import com.amazonaws.auth.*;
+import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRepositories;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
+import com.amazonaws.client.builder.AwsClientBuilder;
+
+import com.amazonaws.auth.AWSCredentials;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+
+@Configuration
+@EnableDynamoDBRepositories(basePackages = "iti")
+public class DynamoDBConfig {
+
+    @Value("${amazon.aws.accesskey}")
+    private String awsAccessKey;
+
+    @Value("${amazon.aws.secretkey}")
+    private String awsSecretKey;
+
+    @Value("${amazon.dynamodb.endpoint:}")
+    private String amazonDynamoDBEndpoint;
+
+    public AWSCredentialsProvider amazonAWSCredentialsProvider() {
+        return new AWSStaticCredentialsProvider(awsCredentials());
+    }
+
+    @Bean
+    public AWSCredentials awsCredentials() {
+        return new BasicAWSCredentials(awsAccessKey, awsSecretKey);
+    }
+
+    @Bean
+    public AmazonDynamoDB amazonDynamoDB(AWSCredentials awsCredentials) {
+
+        AmazonDynamoDB amazonDynamoDB;
+        amazonDynamoDB =  AmazonDynamoDBClientBuilder.standard()
+                .withCredentials(amazonAWSCredentialsProvider())
+                .withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(amazonDynamoDBEndpoint, "AP_SOUTHEAST_2"))
+                .build();
+        return amazonDynamoDB;
+    }
+}
