+++ src/main/java/iti/db/repository/ProductRepositoryInterface.java
@@ -1,0 +1,14 @@
+package iti.db.repository;
+
+import iti.model.Product;
+import org.socialsignin.spring.data.dynamodb.repository.EnableScan;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+
+@EnableScan
+public interface ProductRepositoryInterface extends CrudRepository<Product, String> {
+
+    Product saveProduct(Product product);
+    List<Product> findAllProducts();
+}
