--- src/main/resources/application.properties
@@ -1,3 +1,0 @@
-amazon.dynamodb.endpoint = http://localhost:8000/
-amazon.aws.accesskey = key
-amazon.aws.secretkey = key2
